Store integers as 64-bit
Doubles are the same size, so I think it's fine to also allow integers
to be 64-bit. This allows the user to store pointers on amd64, e.g.
as ID for asynchronous events.

+++ include/spdr/spdr-private.h
@@ -1,6 +1,8 @@
 #ifndef UU_SPDR_PRIVATE_H
 #define UU_SPDR_PRIVATE_H
 
+#include "pstdint.h"
+
 /**
  * Context for the library
  */
@@ -23,13 +25,13 @@ struct SPDR_Event_Arg {
         const char *key;
         enum SPDR_Event_Arg_Type type;
         union {
-                int i;
+                int64_t i;
                 double d;
                 const char *str;
         } value;
 };
 
-extern struct SPDR_Event_Arg uu_spdr_arg_make_int(const char *key, int value);
+extern struct SPDR_Event_Arg uu_spdr_arg_make_int(const char *key, int64_t value);
 extern struct SPDR_Event_Arg uu_spdr_arg_make_double(const char *key,
                                                      double value);
 extern struct SPDR_Event_Arg uu_spdr_arg_make_str(const char *key,
